[nextest-runner] move RunStats and related types to reporter
This is part of the external interface, so it should live in
`reporter/events.rs`.
