+++ cargo-nextest/src/dispatch.rs
@@ -31,11 +31,11 @@ use nextest_runner::{
     platform::{BuildPlatforms, HostPlatform, PlatformLibdir, TargetPlatform},
     redact::Redactor,
     reporter::{
-        highlight_end, structured, FinalStatusLevel, StatusLevel, TestOutputDisplay,
-        TestOutputErrorSlice, TestReporterBuilder,
+        highlight_end, structured, FinalRunStats, FinalStatusLevel, RunStatsFailureKind,
+        StatusLevel, TestOutputDisplay, TestOutputErrorSlice, TestReporterBuilder,
     },
     reuse_build::{archive_to_file, ArchiveReporter, PathMapper, ReuseBuildInfo},
-    runner::{configure_handle_inheritance, FinalRunStats, RunStatsFailureKind, TestRunnerBuilder},
+    runner::{configure_handle_inheritance, TestRunnerBuilder},
     show_config::{ShowNextestVersion, ShowTestGroupSettings, ShowTestGroups, ShowTestGroupsMode},
     signal::SignalHandlerKind,
     target_runner::{PlatformRunner, TargetRunner},
