+++ nextest-runner/tests/integration/target_runner.rs
@@ -11,7 +11,8 @@ use nextest_runner::{
     double_spawn::DoubleSpawnInfo,
     input::InputHandlerKind,
     platform::{BuildPlatforms, HostPlatform, PlatformLibdir, TargetPlatform},
-    runner::{FinalRunStats, RunStatsFailureKind, TestRunnerBuilder},
+    reporter::{FinalRunStats, RunStatsFailureKind},
+    runner::TestRunnerBuilder,
     signal::SignalHandlerKind,
     target_runner::{PlatformRunner, TargetRunner},
     test_filter::{RunIgnored, TestFilterBuilder},
